Create Common Assessment 1 (b)(c)
(b) It was felt that the list is not a very efficient data structure, and each case should be in a dictionary.
The name of the case should be the key and there should be a list with the colour and the phone
supported

(c)(i) Now that the data is in a dictionary, create a binary search to quickly find a case. It should output
the name of the items and the associated list or a suitable message if no case is found.